Added white versions of the meshes, so that color changes make sense. FInished loading Material from binary. Fixed bugs with MaterialBuilder not properly indexing things (because Lua is dumb and uses 1 based indices) Removed shading_group from the maya mesh exporter (it is only an internal value that is used by maya) altered the material on all the game objects to use the new materials. MaterialBuilder now appends the game data dir to the front of the effect file, just as the effect appends it to the shader files.

+++ Code/Engine/Graphics/Direct3D/Effect.d3d.cpp
@@ -181,7 +181,7 @@ namespace Lame
 
 	bool Effect::SetConstant(const Shader &i_shader, const ConstantHandle &i_constant, const float *i_val, const size_t &i_val_count)
 	{
-		HRESULT result = get_constant_table(i_shader)->SetFloatArray(context->get_direct3dDevice(), i_constant, i_val, i_val_count);
+		HRESULT result = get_constant_table(i_shader)->SetFloatArray(context->get_direct3dDevice(), i_constant, i_val, static_cast<UINT>(i_val_count));
 		if (!SUCCEEDED(result))
 		{
 			System::UserOutput::Display("DirectX failed to set a constant uniform value.");
